Quiz fixed and updated controllers

+++ routes/web.php
@@ -28,14 +28,14 @@ Route::group(['middleware' => ['auth']], function () {
     Route::get('/messages', 'Admin\MessageController@index');
     Route::get('/settings', 'Admin\SettingController@index')->name('settings.index');
     Route::post('/settings/{setting}', 'Admin\SettingController@update')->name('settings.update');
-    Route::get('/dashboard', 'HomeController@index')->name('home');
+    Route::get('/dashboard', 'Admin\HomeController@index')->name('home');
 
     Route::resource('team-members', 'Admin\TeamMemberController')->except('show');
-    Route::resource('sliders', 'SliderController');
-    Route::resource('books', 'BookController');
-    Route::resource('news-updates', 'NewsController');
-    Route::resource('publications', 'PublicationController');
-    Route::resource('blogs', 'BlogController');
+    Route::resource('sliders', 'Admin\SliderController');
+    Route::resource('books', 'Admin\BookController');
+    Route::resource('news-updates', 'Admin\NewsController');
+    Route::resource('publications', 'Admin\PublicationController');
+    Route::resource('blogs', 'Admin\BlogController');
 
     Route::get('team-members/{team_member}/highlight', 'Admin\TeamMemberController@highlight')->name('team-members.highlight');
 
@@ -48,41 +48,28 @@ Route::group(['middleware' => ['auth']], function () {
     Route::get('book/question/delete/{id}','Admin\BookQuestionController@delete')->name('admin.book.question.delete');
 
 
+    Route::resource('galleries', 'Admin\GalleryController')->except('show');
+    Route::resource('videos', 'Admin\VideoController')->except('show');
 
+    Route::post('galleries/{gallery}/slider', 'Admin\GalleryController@setSlider')->name('galleries.slider');
+    Route::get('galleries/{gallery}/photos', 'Admin\GalleryPhotoController@index')->name('gallery-photos.index');
+    Route::post('galleries/{gallery}/photos', 'Admin\GalleryPhotoController@store')->name('gallery-photos.store');
+    Route::delete('galleries-photos/{photo}', 'Admin\GalleryPhotoController@destroy')->name('gallery-photos.delete');
 
-    Route::resource('galleries', 'GalleryController')->except('show');
-    Route::resource('videos', 'VideoController')->except('show');
-
-    Route::post('galleries/{gallery}/slider', 'GalleryController@setSlider')->name('galleries.slider');
-    Route::get('galleries/{gallery}/photos', 'GalleryPhotoController@index')->name('gallery-photos.index');
-    Route::post('galleries/{gallery}/photos', 'GalleryPhotoController@store')->name('gallery-photos.store');
-    Route::delete('galleries-photos/{photo}', 'GalleryPhotoController@destroy')->name('gallery-photos.delete');
-
-
-//    Route::get('/add-gallery', 'GalleryController@addGallery')->name('add-gallery');
-//    Route::get('/new-gallery', 'GalleryController@newGallery')->name('new-gallery');
-//    Route::get('/edit-gallery/{id}', 'GalleryController@editGallery')->name('edit-gallery');
-//    Route::get('/update-gallery/{id}', 'GalleryController@updateGallery')->name('update-gallery');
-//    Route::get('/delete-gallery/{id}', 'GalleryController@deleteGallery')->name('delete-gallery');
-
-
-
-
-    Route::post('/delete-book-image/{id}', 'BookController@deleteBookImage')->name('delete-book-image');
+    Route::post('/delete-book-image/{id}', 'Admin\BookController@deleteBookImage')->name('delete-book-image');
     Route::get('books/{book}/questions','Admin\BookQuestionController@index')->name('book-questions.index');
 
-    Route::get('/bangabandhu-info', 'BangabandhuController@edit')->name('bangabandhu.edit');
-    Route::patch('/bangabandhu-info', 'BangabandhuController@update')->name('bangabandhu.update');
+    Route::get('/bangabandhu-info', 'Admin\BangabandhuController@edit')->name('bangabandhu.edit');
+    Route::patch('/bangabandhu-info', 'Admin\BangabandhuController@update')->name('bangabandhu.update');
 
-    Route::get('contacts/edit', 'ContractController@edit')->name('contacts.edit');
-    Route::post('contacts/{contact}', 'ContractController@update')->name('contacts.update');
+    Route::get('contacts/edit', 'Admin\ContactController@edit')->name('contacts.edit');
+    Route::post('contacts/{contact}', 'Admin\ContactController@update')->name('contacts.update');
 
-    Route::DELETE('/delete-work/{work}', 'WorkController@deleteSubmitWork')->name('delete-work');
-    Route::get('/admin/submitted-work','AdminController@submittedWork')->name('admin.submitted.work');
+    Route::DELETE('/delete-work/{work}', 'Admin\WorkController@deleteSubmitWork')->name('delete-work');
+    Route::get('/admin/submitted-work','Admin\AdminController@submittedWork')->name('admin.submitted.work');
 
-    Route::get('/about/edit', 'AboutController@edit')->name('about.edit');
-    Route::get('/about/{about}/update', 'AboutController@update')->name('about.update');
-
+    Route::get('/about/edit', 'Admin\AboutController@edit')->name('about.edit');
+    Route::get('/about/{about}/update', 'Admin\AboutController@update')->name('about.update');
 });
 
 
@@ -92,25 +79,25 @@ Route::group(['middleware' => ['auth']], function () {
 /*End admin*/
 
 /*Font End*/
-Route::get('/publication-details/{publication}', 'PublicationController@show')->name('publication-details');
-Route::get('add-front','TekasaibdController@front')->name('add-front');
-Route::get('about-us','TekasaibdController@about')->name('about');
-Route::get('contact','TekasaibdController@contact')->name('contact');
-Route::get('privacy','TekasaibdController@privacy')->name('privacy');
-Route::get('tekasaibd','TekasaibdController@tekasaibd')->name('tekasaibd');
-Route::get('bangabandhu','TekasaibdController@bangabandhu')->name('bangabandhu');
-Route::get('amaderkotha','TekasaibdController@amaderkotha')->name('amaderkotha');
-Route::get('blog','TekasaibdController@blog')->name('blog');
-Route::get('/blog-details/{blog}', 'TekasaibdController@detailsBlog')->name('blog-details');
-Route::get('smash-users','SmashUsersController@index')->name('users-smash');
-Route::post('/send-email', 'TekasaibdController@sendEmail')->name('send-email');
-Route::get('/book/details/{book}', 'TekasaibdController@bookDetails')->name('book.details');
-Route::get('/submit-work', 'WorkController@addWork')->name('submit-work');
-Route::post('/new-work','WorkController@newWork')->name('new-work');
+Route::get('/publication-details/{publication}', 'Admin\PublicationController@show')->name('publication-details');
+Route::get('add-front','Website\WebsiteController@front')->name('add-front');
+Route::get('about-us','Website\WebsiteController@about')->name('about');
+Route::get('contact','Website\WebsiteController@contact')->name('contact');
+Route::get('privacy','Website\WebsiteController@privacy')->name('privacy');
+Route::get('tekasaibd','Website\WebsiteController@about')->name('tekasaibd');
+Route::get('bangabandhu','Website\WebsiteController@bangabandhu')->name('bangabandhu');
+Route::get('amaderkotha','Website\WebsiteController@amaderkotha')->name('amaderkotha');
+Route::get('blog','Website\WebsiteController@blog')->name('blog');
+Route::get('/blog-details/{blog}', 'Website\WebsiteController@detailsBlog')->name('blog-details');
+Route::get('smash-users','Admin\SmashUsersController@index')->name('users-smash');
+Route::post('/send-email', 'Website\WebsiteController@sendEmail')->name('send-email');
+Route::get('/book/details/{book}', 'Website\WebsiteController@bookDetails')->name('book.details');
+Route::get('/submit-work', 'Admin\WorkController@addWork')->name('submit-work');
+Route::post('/new-work','Admin\WorkController@newWork')->name('new-work');
 Route::get('/team-member/{member}/show','Admin\TeamMemberController@show')->name('team-members.show');
-Route::get('/gallery','TekasaibdController@gallery')->name('gallery.list');
-Route::get('/gallery/{gallery}','GalleryController@show')->name('galleries.show');
-Route::get('/gallery/{videos}','VideoPreviewController@show')->name('gallery-videos.show');
+Route::get('/gallery','Website\WebsiteController@gallery')->name('gallery.list');
+Route::get('/gallery/{gallery}','Admin\GalleryController@show')->name('galleries.show');
+Route::get('/gallery/{videos}','Admin\VideoPreviewController@show')->name('gallery-videos.show');
 
 Route::view('under-construction','front.dump.construction');
 Route::view('terms-and-conditions','front.dump.terms-and-conditions');
